+++ Vostok.ClusterClient.Core/Misc/LastAcceptedResponseSelector.cs
@@ -9,7 +9,8 @@ namespace Vostok.Clusterclient.Core.Misc
     /// Represents a response selector which works using following priority system:
     /// <list type="number">
     /// <item><description>If there are no results at all, it returns <c>null</c>.</description></item>
-    /// <item><description>If there are any results with <see cref="ResponseVerdict.Accept"/> verdict, it returns the last of them.</description></item>
+    /// <item><description>If there are any results with <see cref="ResponseVerdict.Accept"/> verdict, it returns the last of them.
+    /// Results without <see cref="HeaderNames.UnreliableResponse"/> header will be more preferable than with it.</description></item>
     /// <item><description>If there are any results with response code other than <see cref="ResponseCode.Unknown"/>, it returns the last of them.</description></item>
     /// <item><description>As a last resort, it just returns response of last result in the list.</description></item>
     /// </list>
@@ -22,6 +23,7 @@ namespace Vostok.Clusterclient.Core.Misc
             GetLastAcceptedResponse(results) ?? GetLastKnownResponse(results) ?? GetLastResponse(results);
 
         private static Response GetLastAcceptedResponse(IList<ReplicaResult> results) =>
+            results.LastOrDefault(result => result.Verdict == ResponseVerdict.Accept && result.Response.Headers[HeaderNames.UnreliableResponse] == null)?.Response ??
             results.LastOrDefault(result => result.Verdict == ResponseVerdict.Accept)?.Response;
 
         private static Response GetLastKnownResponse(IList<ReplicaResult> results) =>